requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==1.4.52
python-dotenv==1.0.1
aiofiles==23.2.1
python-multipart==0.0.9
python-telegram-bot==21.1.1
httpx==0.27.0
